Rename/색상만 다른 중복 Github Icon이 있어 파일명 변경
